backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pydantic[email]
sqlalchemy
psycopg2-binary
PyJWT
passlib
bcrypt
python-jose
cryptography
jose
python-dateutil
httpx
pytz
fastapi_mail